Harden control plane: close fe80 hijack, atomic nft rebuild, bounded copy queue
Review findings, by priority:

- host_ips now includes link-local (was global-only): a guest could claim
  the host's own fe80 (learnable) and have unicast-to-host-LL traffic
  (router NUD probes, unicast RA) demuxed away from the host. LL addrs now
  feed the kernel host4/host6 accept sets and the reconcile host-wins skip.
  Keepalive probe sender skips LL v4.

- nft rebuild is now ONE atomic batch (DELTABLE + table + sets + chains +
  all rules): the old delete-then-rebuild-per-rule flow left a window in
  direct mode where guest src macs leaked out up0 (case 13). Same-batch
  set references resolve via NFTA_LOOKUP_SET_ID / NFTA_DYNSET_SET_ID.
  Element re-flash stays in follow-up transactions (fail-safe: empty sets
  still normalize src via the else row).

- copy queue (readers -> core) is bounded (1024) + lossy try_send: kernel
  copies are already lossy; an unbounded queue let a guest cycling src
  addresses balloon memory while reconcile (netlink RTTs) lagged.

- netlink monitor death (e.g. ENOBUFS) is now detected and the monitor
  rebuilt + resynced instead of silently disabling topology tracking.

- SIGINT stream hoisted out of the select loop (per-iteration ctrl_c()
  could miss a signal); aging timer uses MissedTickBehavior::Delay so a
  stalled loop doesn't burst-flush and halve the effective timeout; a
  failed init is retried on the aging tick, not only on netlink events;
  nft-ND copies learn BEFORE reinject so the discovery-dup dedup sees the
  entry; dump_set_elems propagates recv errors instead of returning a
  partial alive-set that would evict live entries.

- Cleanups: clippy clean (vec_init_then_push, casts, parens); stale docs
  (LearnResult, vmroute defaults); explicit 3-byte pad in BPF cfg struct;
  bpf_obj_parses test now fails on object-format errors.

Validated: cargo clippy --all-targets clean, 25 unit tests pass, smoke
4/4, full matrix ENV=linux 6 configs (direct/fwd/fwd-with-offload x
nft/ebpf) 16/16 each — ALL CONFIGS PASSED.

+++ src/afpacket.rs
@@ -24,7 +24,7 @@ impl Injector {
             libc::socket(
                 libc::AF_PACKET,
                 libc::SOCK_RAW | libc::SOCK_CLOEXEC,
-                (ETH_P_ALL as u16).to_be() as i32,
+                ETH_P_ALL.to_be() as i32,
             )
         };
         if raw < 0 {
@@ -45,7 +45,7 @@ impl Injector {
         }
         let sa = libc::sockaddr_ll {
             sll_family: libc::AF_PACKET as u16,
-            sll_protocol: (ETH_P_ALL as u16).to_be(),
+            sll_protocol: ETH_P_ALL.to_be(),
             sll_ifindex: self.ifindex as i32,
             sll_hatype: 0,
             sll_pkttype: 0,
